всё-таки выкорчевал маргины, заменив их на систему layout/rendering transforms, работающую по принципу опциональной модификации границ дочерних объектов - теперь и оператива не хавается по хуйне, и лишние расчёты с нулевыми маргинами не производятся, и трансформы ебически гибкие стали

+++ src/YOBA/Core/Margin.cpp
@@ -3,43 +3,43 @@
 namespace YOBA {
 	Margin Margin::zero { 0 };
 
-	void Margin::setBottom(const int32_t value) {
+	void Margin::setBottom(const uint16_t value) {
 		_bottom = value;
 	}
 
-	int32_t Margin::getBottom() const {
+	uint16_t Margin::getBottom() const {
 		return _bottom;
 	}
 
-	void Margin::setRight(const int32_t value) {
+	void Margin::setRight(const uint16_t value) {
 		_right = value;
 	}
 
-	int32_t Margin::getRight() const {
+	uint16_t Margin::getRight() const {
 		return _right;
 	}
 
-	void Margin::setTop(const int32_t value) {
+	void Margin::setTop(const uint16_t value) {
 		_top = value;
 	}
 
-	int32_t Margin::getTop() const {
+	uint16_t Margin::getTop() const {
 		return _top;
 	}
 
-	void Margin::setLeft(const int32_t value) {
+	void Margin::setLeft(const uint16_t value) {
 		_left = value;
 	}
 
-	int32_t Margin::getLeft() const {
+	uint16_t Margin::getLeft() const {
 		return _left;
 	}
 
-	int32_t Margin::getHorizontal() const {
+	uint16_t Margin::getHorizontal() const {
 		return _left + _right;
 	}
 
-	int32_t Margin::getVertical() const {
+	uint16_t Margin::getVertical() const {
 		return _top + _bottom;
 	}
