+++ src/YOBA/Core/Margin.hpp
@@ -5,15 +5,15 @@
 namespace YOBA {
 	class Margin {
 		public:
-			constexpr Margin(const int32_t left, const int32_t top, const int32_t right, const int32_t bottom) : _left(left), _top(top), _right(right), _bottom(bottom) {
+			constexpr Margin(const uint16_t left, const uint16_t top, const uint16_t right, const uint16_t bottom) : _left(left), _top(top), _right(right), _bottom(bottom) {
 
 			}
 
-			constexpr Margin(const int32_t horizontal, const int32_t vertical) : _left(horizontal), _top(vertical), _right(horizontal), _bottom(vertical) {
+			constexpr Margin(const uint16_t horizontal, const uint16_t vertical) : _left(horizontal), _top(vertical), _right(horizontal), _bottom(vertical) {
 
 			}
 
-			constexpr Margin(const int32_t uniformValue) : _left(uniformValue), _top(uniformValue), _right(uniformValue), _bottom(uniformValue) {
+			constexpr Margin(const uint16_t uniformValue) : _left(uniformValue), _top(uniformValue), _right(uniformValue), _bottom(uniformValue) {
 
 			}
 
@@ -23,28 +23,28 @@ namespace YOBA {
 
 			static Margin zero;
 
-			int32_t getLeft() const;
-			void setLeft(int32_t value);
+			uint16_t getLeft() const;
+			void setLeft(uint16_t value);
 
-			int32_t getTop() const;
-			void setTop(int32_t value);
+			uint16_t getTop() const;
+			void setTop(uint16_t value);
 
-			int32_t getRight() const;
-			void setRight(int32_t value);
+			uint16_t getRight() const;
+			void setRight(uint16_t value);
 
-			int32_t getBottom() const;
-			void setBottom(int32_t value);
+			uint16_t getBottom() const;
+			void setBottom(uint16_t value);
 
-			int32_t getHorizontal() const;
-			int32_t getVertical() const;
+			uint16_t getHorizontal() const;
+			uint16_t getVertical() const;
 
 			bool operator==(const Margin& rhs) const;
 			bool operator!=(const Margin& rhs) const;
 
 		private:
-			int32_t _left;
-			int32_t _top;
-			int32_t _right;
-			int32_t _bottom;
+			uint16_t _left;
+			uint16_t _top;
+			uint16_t _right;
+			uint16_t _bottom;
 	};
 }
